Replace function memoizer
The previous function memoizer was short and nice, but kind of hacky,
and it was impossible to extend it to inherit the docstring and
metadata, as well as accept nonhashable arguments in any decent manner.
The new function memoizer is adapted from the method memoizer, and even
contains a clear_cache method.