Make classes with memoized methods picklable
Memoized methods used to introduce instance variables for caches, which
would contain functions not declared on the module level. This made the
instances unpicklable. This is fixed by not caching the bound methods
themselves, and using the function names instead of function objects in
the keys to the return value cache.